Moved search box to header
