+++ netlify/functions/constants.ts
@@ -1,5 +1,5 @@
 /** Movie DB API key */
-export const apiKey = process.env.SEARCHOVIE_API_KEY;
+export const apiKey = process.env.SEARCHOVIE_API_KEY || 'YOUR_API_KEY_HERE';
 
 /** Base URL of movie db API */
 export const apiBase = 'https://api.themoviedb.org/3/';
