+++ package.json
@@ -14,7 +14,8 @@
     "@mui/material": "^5.5.2",
     "axios": "^0.26.1",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^5.3.0"
   },
   "devDependencies": {
     "@netlify/functions": "^1.0.0",
@@ -22,6 +23,7 @@
     "@types/lodash": "^4.14.180",
     "@types/react": "^17.0.33",
     "@types/react-dom": "^17.0.10",
+    "@types/react-router-dom": "^5.3.3",
     "@vitejs/plugin-react": "^1.0.7",
     "autoprefixer": "^10.4.4",
     "lodash": "^4.17.21",
