+++ src/MovieDetails.tsx
@@ -1,0 +1,5 @@
+function MovieDetails() {
+  return <div>Movie Details</div>
+};
+
+export default MovieDetails;
